pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "argus-for-athena"
version = "1.0.0"
description = "A local DBMS CLI for AWS Athena with automatic workgroup routing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "boto3>=1.34.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sse-starlette>=1.8.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pyarrow>=14.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "awscrt>=0.19.0",
    "mangum>=0.17.0",
    "PyJWT[cryptography]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "moto[athena,glue,s3]>=5.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
argus = "argus.cli.main:app"
argus-ui = "argus.api.app:run_server"

[tool.setuptools.packages.find]
where = ["src"]
